안쓰는 자바 프래그먼트 대대 정리 -->
해당 프래그먼트
HomeTag1Adapter.java
HomeTag1Item.java
HomeTag1ViewHolder.java

HomeTag2Adapter.java
HomeTag2Item.java
HomeTag2ViewHolder.java

ListFragment.java
ListSearchAdapter.java
ListSearchFragment.java
ListSearchItem.java
ListSearchViewHolder.java

ReviewSearchAdapter.java
ReviewSearchFragment.java
ReviewSearchItem.java
ReviewSearchViewHolder.java

나머지는 import 변경
