+++ app/src/main/java/com/example/wmc/ListCafeList/ListCafeListViewHolder.java
@@ -3,13 +3,11 @@ package com.example.wmc.ListCafeList;
 import android.view.View;
 import android.widget.CheckBox;
 import android.widget.ImageView;
-import android.widget.RatingBar;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.wmc.HomeTag1.HomeTag1Adapter;
 import com.example.wmc.R;
 
 public class ListCafeListViewHolder extends RecyclerView.ViewHolder{
