+++ app/src/main/java/com/example/wmc/MainActivity.java
@@ -1,36 +1,17 @@
 package com.example.wmc;
 
-import android.content.ClipData;
-import android.content.Intent;
-import android.net.Uri;
 import android.os.Bundle;
-import android.provider.MediaStore;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.Toast;
 
-import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
 
 import com.example.wmc.databinding.ActivityMainBinding;
-import com.example.wmc.ui.Fragment.HomeFragment;
-import com.example.wmc.ui.Fragment.ListFragment;
-import com.example.wmc.ui.Fragment.MyPageFragment;
-import com.example.wmc.ui.Fragment.ReviewFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-
-import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
