+++ app/src/main/java/com/example/wmc/a_No_USE_Files/ListSearchItem.java
@@ -1,4 +1,4 @@
-package com.example.wmc.ListSearch;
+package com.example.wmc.a_No_USE_Files;
 
 public class ListSearchItem {
