+++ app/src/main/java/com/example/wmc/a_No_USE_Files/ListSearchViewHolder.java
@@ -1,14 +1,12 @@
-package com.example.wmc.ListSearch;
+package com.example.wmc.a_No_USE_Files;
 
 import android.view.View;
 import android.widget.Button;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.wmc.CafeDetailMore.CafeDetailMoreAdapter;
 import com.example.wmc.R;
 
 public class ListSearchViewHolder extends RecyclerView.ViewHolder{
