+++ app/src/main/java/com/example/wmc/ui/Fragment/HomeFragment.java
@@ -14,12 +14,10 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.RequiresApi;
 import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import androidx.viewpager.widget.ViewPager;
 
 import com.android.volley.Cache;
 import com.android.volley.Network;
@@ -31,19 +29,11 @@ import com.android.volley.toolbox.BasicNetwork;
 import com.android.volley.toolbox.DiskBasedCache;
 import com.android.volley.toolbox.HurlStack;
 import com.android.volley.toolbox.StringRequest;
-import com.example.wmc.CafeDetail.CafeDetailAdapter;
-import com.example.wmc.CafeDetail.CafeDetailItem;
-import com.example.wmc.CafeDetailImageViewPager.CafeDetailRatingItem;
-import com.example.wmc.CafeDetailImageViewPager.CafeDetailRatingViewPagerAdapter;
 import com.example.wmc.HomeFavorite.HomeFavoriteAdapter;
 import com.example.wmc.HomeFavorite.HomeFavoriteItem;
-import com.example.wmc.HomeTag1.HomeTag1Adapter;
-import com.example.wmc.HomeTag1.HomeTag1Item;
 import com.example.wmc.HomeTag1ViewPager.HomeTag1ViewPager;
 import com.example.wmc.HomeTag1ViewPager.HomeTag1ViewPagerAdapter;
 import com.example.wmc.HomeTag1ViewPager.HomeTag1ViewPagerItem;
-import com.example.wmc.HomeTag2.HomeTag2Adapter;
-import com.example.wmc.HomeTag2.HomeTag2Item;
 import com.example.wmc.MainActivity;
 import com.example.wmc.HomeTag2ViewPager.HomeTag2ViewPager;
 import com.example.wmc.HomeTag2ViewPager.HomeTag2ViewPagerAdapter;
