+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/navigation_list"
-        android:name="com.example.wmc.ui.Fragment.ListFragment"
+        android:name="com.example.wmc.a_No_USE_Files.ListFragment"
         android:label="@string/title_list"
         tools:layout="@layout/fragment_list" >
         <action
@@ -94,7 +94,7 @@
 
     <fragment
         android:id="@+id/navigation_list_search"
-        android:name="com.example.wmc.ui.Fragment.ListSearchFragment"
+        android:name="com.example.wmc.a_No_USE_Files.ListSearchFragment"
         android:label="@string/list_search"
         tools:layout="@layout/fragment_list_search" >
 
@@ -192,7 +192,7 @@
 
     <fragment
         android:id="@+id/navigation_review_search"
-        android:name="com.example.wmc.ui.Fragment.ReviewSearchFragment"
+        android:name="com.example.wmc.a_No_USE_Files.ReviewSearchFragment"
         android:label="@string/review_search"
         tools:layout="@layout/fragment_list_search" >
